// apps/frontend/src/api/generated/model/quizFeedbackSummaryDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * QuizDock API
 * API REST du builder de quiz et des restitutions
 * OpenAPI spec version: 0.1.0
 */
import type { QuizFeedbackSummaryDtoItemsItem } from './quizFeedbackSummaryDtoItemsItem';

export interface QuizFeedbackSummaryDto {
  /**
     * @minimum -9007199254740991
     * @maximum 9007199254740991
     */
  count: number;
  average: number;
  /**
     * @minItems 5
     * @maxItems 5
     * @items.minimum -9007199254740991
     * @items.maximum 9007199254740991
     */
  distribution: number[];
  items: QuizFeedbackSummaryDtoItemsItem[];
  /**
     * @minimum -9007199254740991
     * @maximum 9007199254740991
     */
  page: number;
  /**
     * @minimum -9007199254740991
     * @maximum 9007199254740991
     */
  pageSize: number;
  /**
     * @minimum -9007199254740991
     * @maximum 9007199254740991
     */
  total: number;
}